add cli flags
