--- tests/old/by_ref.ron
@@ -1,159 +1,0 @@
-Test(
-    name: "by_ref",
-    funcs: [
-        (
-            name: "pass_ref",
-            conventions: [All],
-            inputs: [Ref(Struct("MyStruct", [Int(c_uint8_t(0xf1)), Float(c_double(1234.23))]))],
-            output: None,
-        ),
-        (
-            name: "return_ref",
-            conventions: [All],
-            inputs: [],
-            output: Some(Ref(Struct("MyStruct", [Int(c_uint8_t(0xf1)), Float(c_double(1234.23))]))),
-        ),
-        (
-            name: "both_ref",
-            conventions: [All],
-            inputs: [Ref(Struct("MyStruct", [Int(c_uint8_t(0x12)), Float(c_double(0.5891))]))],
-            output: Some(Ref(Struct("MyStruct", [Int(c_uint8_t(0xf1)), Float(c_double(1234.23))]))),
-        ),
-        (
-            name: "pass_array_ref",
-            conventions: [All],
-            inputs: [Ref(Array([
-                Int(c_uint32_t(0x1234_567f)),
-                Int(c_uint32_t(0xae34_5e7a)),
-                Int(c_uint32_t(0xb234_e678)),
-                Int(c_uint32_t(0xa2e4_462f)),
-                Int(c_uint32_t(0x1204_5608)),
-                Int(c_uint32_t(0x09b2_7421)),
-                Int(c_uint32_t(0xf5e1_e972)),
-            ]))],
-            output: None,
-        ),
-        (
-            name: "return_array_ref",
-            conventions: [All],
-            inputs: [],
-            output: Some(Ref(Array([
-                Int(c_uint32_t(0x1234_567f)),
-                Int(c_uint32_t(0xae34_5e7a)),
-                Int(c_uint32_t(0xb234_e678)),
-                Int(c_uint32_t(0xa2e4_462f)),
-                Int(c_uint32_t(0x1204_5608)),
-                Int(c_uint32_t(0x09b2_7421)),
-                Int(c_uint32_t(0xf5e1_e972)),
-            ]))),
-        ),
-        (
-            name: "both_array_ref",
-            conventions: [All],
-            inputs: [Ref(Array([
-                Int(c_uint32_t(0xe13e_362f)),
-                Int(c_uint32_t(0xaf34_5e1a)),
-                Int(c_uint32_t(0xb230_e178)),
-                Int(c_uint32_t(0xa2e4_16ef)),
-                Int(c_uint32_t(0x1401_540e)),
-                Int(c_uint32_t(0x3912_7423)),
-                Int(c_uint32_t(0xf131_3972)),
-            ]))],
-            output: Some(Ref(Array([
-                Int(c_uint32_t(0x1234_567f)),
-                Int(c_uint32_t(0xae34_5e7a)),
-                Int(c_uint32_t(0xb234_e678)),
-                Int(c_uint32_t(0xa2e4_462f)),
-                Int(c_uint32_t(0x1204_5608)),
-                Int(c_uint32_t(0x09b2_7421)),
-                Int(c_uint32_t(0xf5e1_e972)),
-            ]))),
-        ),
-        (
-            name: "multi_dimensional_arrayed",
-            conventions: [All],
-            inputs: [Ref(Array([
-                Array([Int(c_uint32_t(0xe13e_362f)), Int(c_uint32_t(0xaf34_5e1a))]),
-                Array([Int(c_uint32_t(0xb230_e178)), Int(c_uint32_t(0xa2e4_16ef))]),
-                Array([Int(c_uint32_t(0x1401_540e)), Int(c_uint32_t(0x3912_7423))]),
-            ]))],
-            output: Some(
-                Ref(Array([
-                    Array([Int(c_uint8_t(0xae)), Int(c_uint8_t(0xbc))]),
-                    Array([Int(c_uint8_t(0x21)), Int(c_uint8_t(0x3f))]),
-                    Array([Int(c_uint8_t(0x00)), Int(c_uint8_t(0x12))]),
-                ]))
-            ),
-        ),
-        (
-            name: "array_of_struct_ref",
-            conventions: [All],
-            inputs: [Ref(Array([
-                Struct("MyStruct", [Int(c_uint8_t(0xf1)), Float(c_double(1234.23))]), 
-                Struct("MyStruct", [Int(c_uint8_t(0x1)), Float(c_double(0.23))]),
-            ]))],
-            output: Some(
-                Ref(Array([
-                    Struct("MyStruct", [Int(c_uint8_t(0xae)), Float(c_double(0.0002343))]), 
-                    Struct("MyStruct", [Int(c_uint8_t(0x23)), Float(c_double(134123123.232))]),
-                ]))
-            ),
-        ),
-        (
-            name: "array_of_struct_padded_ref",
-            conventions: [All],
-            inputs: [Ref(Array([
-                Struct("MyStruct3", [Float(c_double(1234.23)), Int(c_uint8_t(0xf1))]), 
-                Struct("MyStruct3", [Float(c_double(3.14569)), Int(c_uint8_t(0xa2))]),
-            ]))],
-            output: Some(
-                Ref(Array([
-                    Struct("MyStruct3", [Float(c_double(4.123124)), Int(c_uint8_t(0x34))]), 
-                    Struct("MyStruct3", [Float(c_double(0.0023)), Int(c_uint8_t(0xeb))]),
-                ]))
-            ),
-        ),
-        (
-            name: "complex_arrayed_ref",
-            conventions: [All],
-            inputs: [Ref(Struct("MyComplexArrayed", [
-                Int(c_uint8_t(0xaf)),
-                Array([
-                    Struct("MyComplex", [
-                        Struct("MyStruct", [Int(c_uint8_t(0xf1)), Float(c_double(1234.23))]), 
-                        Struct("MyStruct2", [Float(c_float(81344789.12)), Int(c_uint16_t(0x5678))]),
-                    ]),
-                    Struct("MyComplex", [
-                        Struct("MyStruct", [Int(c_uint8_t(0xf1)), Float(c_double(91.00001))]), 
-                        Struct("MyStruct2", [Float(c_float(3.245)), Int(c_uint16_t(0x1234))]),
-                    ]),
-                    Struct("MyComplex", [
-                        Struct("MyStruct", [Int(c_uint8_t(0xf1)), Float(c_double(23.459))]), 
-                        Struct("MyStruct2", [Float(c_float(0.1234)), Int(c_uint16_t(0xaef3))]),
-                    ]),
-                ]),
-                Int(c_uint64_t(0x3ae1_2345_2a01_318b)),
-            ]))],
-            output: Some(
-                Ref(Struct("MyComplexArrayed", [
-                    Int(c_uint8_t(0x12)),
-                    Array([
-                        Struct("MyComplex", [
-                            Struct("MyStruct", [Int(c_uint8_t(0xa2)), Float(c_double(12.83123432))]), 
-                            Struct("MyStruct2", [Float(c_float(81344789.12)), Int(c_uint16_t(0x8672))]),
-                        ]),
-                        Struct("MyComplex", [
-                            Struct("MyStruct", [Int(c_uint8_t(0x1f)), Float(c_double(1234589))]), 
-                            Struct("MyStruct2", [Float(c_float(133.346)), Int(c_uint16_t(0x9224))]),
-                        ]),
-                        Struct("MyComplex", [
-                            Struct("MyStruct", [Int(c_uint8_t(0x49)), Float(c_double(123823.4597))]), 
-                            Struct("MyStruct2", [Float(c_float(0.01338)), Int(c_uint16_t(0x12f4))]),
-                        ]),
-                    ]),
-                    Int(c_uint64_t(0x12e1_4395_7a08_a131)),
-                ]))
-            )
-        ),
-    ]
-)
